Reloaded:  * Fix build for windows. pthread.h #includes config.h and we used configBis.h, so fix that.  * Reorder some includes  * Fix a typo in synchro.cpp

+++ Reloaded/src/core/render.h
@@ -432,7 +432,7 @@ class Renderer
 
 	int mapGAEntryToRGB(char pal)
 	{
-		return _renderFunc->MapRGB(_colours[pal]);
+		return _renderFunc->MapRGB(_colours[(int)pal]);
 	}
 
 	//! Render CPC memory
